requirements.txt
pandas
numpy
pyyaml
torchsummary
opencv-python
opencv-contrib-python
pillow
torchvision
matplotlib
tqdm
albumentations
pydantic